update toggle interactive example
