+++ packages/styleguide/src/lib/Atoms/Toggle/Toggle.mdx
@@ -71,8 +71,8 @@ If you use voiceover on either of the below, you will hear the difference this m
 
 ## Interactive examples
 
-<Canvas of={ToggleStories.InteractiveInput} />
-<Canvas of={ToggleStories.InteractiveButton} />
+<Canvas of={ToggleStories.InteractiveInputToggle} />
+<Canvas of={ToggleStories.InteractiveButtonToggle} />
 
 
 ## Playground
